refactor: 예약 시간 저장 예외 처리 추가
