+++ src/main/java/roomescape/common/exception/DuplicateException.java
@@ -1,0 +1,7 @@
+package roomescape.common.exception;
+
+public class DuplicateException extends RuntimeException {
+    public DuplicateException(String message) {
+        super(message);
+    }
+}
