+++ src/main/java/roomescape/time/application/exception/DuplicateReservationTimeException.java
@@ -1,0 +1,9 @@
+package roomescape.time.application.exception;
+
+import roomescape.common.exception.DuplicateException;
+
+public class DuplicateReservationTimeException extends DuplicateException {
+    public DuplicateReservationTimeException(String message) {
+        super(message);
+    }
+}
